--- src/components/profile/mapUserDataToRating.tsx
@@ -1,20 +1,0 @@
-import React from 'react';
-import { Rating } from '../../models/dashboard';
-import { UserData } from '../../models/user';
-
-export const mapUserDataToRating = (userData: UserData): Rating[] => {
-    return [
-        {
-            id: userData.id,
-            username: userData.username,
-            points: userData.points,
-            solved: userData.solved,
-            place: userData.place,
-            last_solved: userData.solutions
-                .map((solution) => {
-                    return { date: solution.date, points: solution.task.points };
-                })
-                .concat({ date: userData.date_joined, points: 0 })
-        }
-    ];
-};
