--- src/components/profile/progressStats.tsx
@@ -1,28 +1,0 @@
-import { FaChartLine } from 'react-icons/fa';
-import React from 'react';
-import CardLineChart from '../charts/cardLineChart';
-import { UserData } from '../../models/user';
-import { mapUserDataToRating } from './mapUserDataToRating';
-import { prepareChartData } from '../charts/prepareChartData';
-import { EmptySadFace } from './recentlySolvedListRenderer';
-
-interface ProgressStatsProps {
-    userData: UserData;
-}
-
-const ProgressStats = ({ userData }: ProgressStatsProps) => {
-    const data = prepareChartData(mapUserDataToRating(userData));
-    return (
-        <>
-            <h3 className="flex justify-center items-center space-x-3 ">
-                <span>Postępy w czasie</span>
-                <FaChartLine />
-            </h3>
-            <div className="pt-4 mx-5">
-                {userData.solved ? <CardLineChart ratingData={data} height={'h-96'} /> : <EmptySadFace />}
-            </div>
-        </>
-    );
-};
-
-export default ProgressStats;
