--- src/components/profile/recentlySolved.tsx
@@ -1,24 +1,0 @@
-import { FaCheck } from 'react-icons/fa';
-import React from 'react';
-import RecentlySolvedListRenderer from './recentlySolvedListRenderer';
-import { Solve } from '../../models/solve';
-
-interface RecentlySolvedProps {
-    solves: Solve[];
-    viewingOwn: boolean;
-}
-
-const RecentlySolved = ({ solves, viewingOwn }: RecentlySolvedProps) => {
-    return (
-        <>
-            <h3 className="flex items-center justify-center space-x-2 mb-4 w-full">
-                <span>Ostatnio rozwiązane</span>
-                <FaCheck />
-            </h3>
-
-            <RecentlySolvedListRenderer solves={solves} viewingOwn={viewingOwn} />
-        </>
-    );
-};
-
-export default RecentlySolved;
